--- a/crates/bevy_diagnostic/src/print_diagnostics_plugin.rs
+++ b/crates/bevy_diagnostic/src/print_diagnostics_plugin.rs
@@ -1,7 +1,7 @@
 use super::{Diagnostic, DiagnosticId, Diagnostics};
 use bevy_app::prelude::*;
 use bevy_core::{Time, Timer};
-use bevy_ecs::{IntoQuerySystem, Res, ResMut};
+use bevy_ecs::{IntoSystem, Res, ResMut};
 use std::time::Duration;
 
 /// An App Plugin that prints diagnostics to the console