--- a/crates/bevy_ecs/src/lib.rs
+++ b/crates/bevy_ecs/src/lib.rs
@@ -11,7 +11,7 @@ pub use world::*;
 
 pub mod prelude {
     pub use crate::{
-        resource::{ChangedRes, FromResources, Local, OrRes, Res, ResMut, Resource, Resources},
+        resource::{ChangedRes, FromResources, Local, Res, ResMut, Resource, Resources},
         system::{Commands, IntoSystem, IntoThreadLocalSystem, Query, System},
         world::WorldBuilderSource,
         Added, Bundle, Changed, Component, Entity, Mut, Mutated, Or, QuerySet, Ref, RefMut, With,