--- a/crates/bevy_ecs/src/resource/resources.rs
+++ b/crates/bevy_ecs/src/resource/resources.rs
@@ -1,4 +1,3 @@
-use super::{FetchResource, ResourceQuery};
 use crate::system::SystemId;
 use bevy_hecs::{Archetype, AtomicBorrow, Entity, Ref, RefMut, TypeInfo, TypeState};
 use bevy_utils::HashMap;
@@ -262,29 +261,6 @@ impl Resources {
                 };
                 RefMut::new(&data.archetype, index).ok()
             })
-    }
-
-    pub fn query<Q: ResourceQuery>(&self) -> Option<<Q::Fetch as FetchResource>::Item> {
-        unsafe {
-            if Q::Fetch::is_some(&self, None) {
-                Some(Q::Fetch::get(&self, None))
-            } else {
-                None
-            }
-        }
-    }
-
-    pub fn query_system<Q: ResourceQuery>(
-        &self,
-        id: SystemId,
-    ) -> Option<<Q::Fetch as FetchResource>::Item> {
-        unsafe {
-            if Q::Fetch::is_some(&self, Some(id)) {
-                Some(Q::Fetch::get(&self, Some(id)))
-            } else {
-                None
-            }
-        }
     }
 
     #[inline]