--- a/crates/bevy_ecs/src/system/mod.rs
+++ b/crates/bevy_ecs/src/system/mod.rs
@@ -1,5 +1,6 @@
 mod commands;
 mod into_system;
+mod into_thread_local;
 #[cfg(feature = "profiler")]
 mod profiler;
 mod query;
@@ -8,6 +9,7 @@ mod system;
 
 pub use commands::*;
 pub use into_system::*;
+pub use into_thread_local::*;
 #[cfg(feature = "profiler")]
 pub use profiler::*;
 pub use query::*;