This removes the old "into system" code, which suffered from a variety of problems. Namely:
* system params _must_ follow an arbitrary set of order rules 
* hard to understand
* slow to compile (bevy repo compiles in 1 minutes 13 seconds on my machine)

It replaces it with a new, much simpler approach that has the following benefits:

* system params can be defined in any order
* fast to compile (bevy repo compiles in 48 seconds on my machine)
* much easier to understand

With the old implementation, this code wouldn't compile because it violated the arbitrary "order rules". Now it compiles just fine!
```rust
fn system(a_query: Query<&A>, commands: Commands, x: Res<X>, b_query: Query<&B>) {
}
```

So far these are pretty uncontroversial changes. Now on to the more controversial part: I also removed for-each systems. 

For-each systems were problematic for the following reasons:

1. They increased compile times (I save ~5 seconds in dev builds by removing them)
2. They require a complicated macro, which doesn't fit into the approach used in this pr. We can use the old macro for for-each systems and the new approach for everything else, but that increases maintenance cost / retains an ugly hard to understand macro. 
3. Bevy should generally have "one way to do things". Foreach systems are a _slightly_ more ergonomic way to define a subset of system types. This forces people to make a "design decision" we they shouldn't need to. It also makes examples and tutorials inconsistent according to people's preferences for one or the other. 
4. For-each systems aren't capable of some things, like iterating removed components, controlling iteration, multiple queries at the same time. This means they need to be converted to "query systems" when those features are needed.
5. They have a number of "gotchas" for newcomers that constantly come up in our support forums + confuse newbies:
  * users expect `&mut T` queries to work in foreach systems (ex: `fn system(a: &mut A) {}`). These can't work because we require `Mut<T>` tracking pointers. The equivalent `Query<&mut A>` works because we can return the tracking pointer.
  * The "run on some criteria" bug thats common enough that we cover it in the Bevy Book: https://bevyengine.org/learn/book/getting-started/resources/

## Changes Since Draft

* `SystemParam` can now be derived. We now use this for DrawContext instead of a manual impl. You can use any types that impl SystemParam:

  ```rust
  #[derive(SystemParam)]
  struct MyParams<'a> {
    a: Res<'a, A>, 
    b: ResMut<'a, B>,
    commands: &'a mut Commands,
    #[system_param(ignore)]
    hello: bool // this will use Default
  }

  fn system(my_params: MyParams) {
    println!("{}", system.a);
  }
  ```
* implemented SystemParam for tuples: (SystemParam, SystemParam, SystemParam). Params can now be infinitely nested
* Implement `Or` for system params  
* removed ResourceQueries in favor of SystemParams. also removed UnsafeClone trait and associated impls (score!)
* discovered closure systems work again!
* Removed locks from Commands. We now use `commands: &mut Commands` instead of `mut commands: Commands`. This means Commands can no longer be used in parallel by default. Fortunately ...
* Implemented `SystemParam` for `Arc<Mutex<Commands>>`.
* Simplified ThreadLocalSystem impl

This new approach makes everything so much easier and clearer. I love it :heart: