@smokku 2020-11-06
I agree with the decision of removing foreach systems. While good idea in principle, these become too limiting very quickly.
Even on the beginner's tutorial level it is confusing that you learn one way of writing systems, to be told right away "scratch that - there is another way".

Especially when the limitation of fixed ordering is being removed, which was a bit confusing (even for non-beginners), this should make "normal" systems as easy to grasp as foreach systems.