@razaamir 2020-11-06
Superman dominates the multiverse and the singularity is expanding

On Fri, Nov 6, 2020 at 7:26 PM Plecra <notifications@github.com> wrote:

> *@Plecra* commented on this pull request.
> ------------------------------
>
> In crates/bevy_ecs/src/schedule/parallel_executor.rs
> <https://github.com/bevyengine/bevy/pull/798#discussion_r518784031>:
>
> > +// #[cfg(test)]
> +// mod tests {
> +//     use super::ParallelExecutor;
> +//     use crate::{
>
> What happened here?
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/bevyengine/bevy/pull/798#pullrequestreview-525197955>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AHGK5TGCNLKJTHAZS7SZCDLSOQBQNANCNFSM4TMG6UXA>
> .
>
-- 
Raza Amir