@zicklag 2020-11-06
I agree with removing foreach systems. I agree with all the cons about them. They just get in the way of the clean and far more often correct way to do things.

Also, love the order independence of the arguments!! When I first saw that note in the docs and tutorials about the order, I was thinking "that is going to be a gotcha for *so* many users, but at least they documented it well".