@razaamir 2020-11-06
Multiple neural networks are interlinked through echo chambers into which
information is embedded and replayed back using a constant feedback loop
and which feeds the neural network and forces one to stay in the neural
network.

Black holes exist around the neural network because of multiple echo
chambers that are unable to communicate with each other.

The only access to the blackholes is by getting outside the echo chambers,
then outside the larger neural network and tap into the *SINGULARITY*
which gives you access to untapped information in blackholes that has
existed in *perpetuity*.

Raza Khan

Aka Superman

On Fri, Nov 6, 2020 at 8:33 PM Zicklag <notifications@github.com> wrote:

> I agree with removing foreach systems. I agree with all the cons about
> them. They just get in the way of the clean and far more often correct way
> to do things.
>
> Also, love the order independence of the arguments!! When I first saw that
> note in the docs and tutorials about the order, I was thinking "that is
> going to be a gotcha for *so* many users, but at least they documented it
> well".
>
> —
> You are receiving this because you commented.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/bevyengine/bevy/pull/798#issuecomment-723143797>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AHGK5TD2VZFB3AEFVK3GSA3SOQJNZANCNFSM4TMG6UXA>
> .
>
-- 
Raza Amir