@tigregalis 2020-11-07
Agree with all of the comments so far. I think **parameter order independence** and **having one way of doing things** are much bigger wins for newcomers than having to wrap these parameters in a `Query<>`, and write a `for` loop, are a loss, particularly as:

- not writing my own for-loop for each system is a bit too "magical"
  - it's easier to see what's actually happening with queries i.e. I find all of the entities that has this set of components, and I iterate over these entities, as opposed to (in for-each systems) the ECS will run this function for each entity that has these components
  - for-loops are common enough among programming languages that users should be comfortable with them
  - queries (and then having to do something with those queries) are a familiar enough concept to anyone who has queried a database or data API
- using `Query<T>` means it's a smaller leap to more advanced queries using `Changed`, `Or`, `QuerySet`s, etc.
- for-each systems are almost always mixed in with query systems so there's a bit of context switching within the same codebase or even file, and there's no particularly visible distinction between them either at the function signature or at the point the system gets added to the app
- for-each systems are often rewritten into query systems eventually, and only after you give up trying to do something in a for-each system