@cart 2020-11-07
So far the sentiment to remove for-each systems has been a unanimous "yes please". Thats enough validation for me :smile: 

I'll start polishing this up.