@cart 2020-11-07
So I "fixed" the regression in that it was never really a regression :laughing: 

Turns out my iteration benchmark looked like this:

```rust
for a in &mut query.iter_mut() {
}
```

Which i guess was an artifact of adapting the benchmark to the last round of ecs changes. 

It still works even though we removed the iteration wrapper. However it _tanks_ iteration performance, probably because the indirection prevents rustc from applying optimizations?

So yeah the problem wasn't a problem and now we know that using `&mut query.iter_mut()` instead of `query.iter_mut()` is bad for perf.