@Plecra 2020-11-07
It'd be good to identify exactly which optimisation is being missed, in case there's anything that can be done about it. Composing iterators with `by_ref` is a pretty common pattern and it'd be nice to keep it quick