@cart 2020-11-07
Yeah it would be good to know. But its also worth calling out that "tanking" in this case is 4x a _very_ small number. Adding almost any operation to this benchmark is going to register in a big way (for every ecs framework already operating at this speed).