@memoryruins 2020-11-08
This change definitely makes `bevy` take less time to build, especially on my laptop!

|        	| llvm   	| cranelift 	|
|--------	|--------	|-----------	|
| before 	| 3m 36s 	| 2m 46s    	|
| after  	| 2m 40s 	| 2m 05s    	|

`dev [unoptimized + debuginfo] target(s)`

<details>

<summary>specs / environment</summary>

```
OS: Arch Linux x86_64
Host: ThinkPad T440
Kernel: 5.4.61-1-lts
CPU: Intel i5-4300U (4) @ 2.900GHz
Memory: 7655MiB
```


```toml
# ~/.cargo/config.toml
[target.x86_64-unknown-linux-gnu]
linker = "/usr/bin/clang"
rustflags = ["-Clink-arg=-fuse-ld=lld"]
```

```
active toolchain
----------------
nightly-x86_64-unknown-linux-gnu (default)
rustc 1.49.0-nightly (ffa2e7ae8 2020-10-24)
```

</details>