@jamadazi 2020-11-09
After the new changes, since now I have to use `Arc<Mutex<Commands>>` to have parallel access to `Commands` from many systems, what if I want to use `parking_lot::Mutex` instead of `std::sync::Mutex`? Can this be supported?