@cart 2020-11-09
The impl uses `parking_lot::Mutex` so you should be covered.